Add function object to the evaluator
This requires being able to copy the expressions and statements from the
AST. Implement copy_statement and copy_expression APIs to be able to do
this. When evaluating a function, we need to create a monkey_function_t
object which requires the parameter list and body of the function from
the parser. If we don't copy these, the evaluator will lose them
whenever the parser gets freed.