Now we can evaluate function calls
It is getting hairy as things get complicated. Functions need an
extended environment object to maintain their parameters and global
variables, for this we added an outer field to the environment struct.
But that outer environment might belong to the main program or another
function. Without knowing we cannot free it and if we don't free it, we
leak memory. Therefore to simplify, copy the environment.