Implement parsing of hash literals with tests
Hashing expression objects is tricky, we could implement a hashing API
to hash each type of expression but it could get tedious quickly. A
quick hack right now is to simply use the address of the objects as the
key in the hash table. Till the parser is not free'd we can reference
back the expression object back. The hashing function for pointers is
rather crude and leads to collisions, we may improve that, but we have
separate chining implemented which takes care of it right now.