Create a monkey_hash object with tests
Only strings, integers and bools can be hashed. We already have hash
functions to hash these values. Going on a different path from the one
described in the book here, because the hashing machinery has already
been existing in this implementation and has gone a different direction.
Things are complicated in C because hash table is not a native data
structure. We might consider doing it the book way if things get
complicated.