Add support for arguments to functions
We create a hole on the stack before pushing the new frame for the
function being called. The hole gets used for the local variables of the
function. As well as before calling the function we push all the
arguments to the function onto the stack and the function reads them
similar to how it would read the local variables. We save the previous
stack pointer as the field bp (for base pointer in the new frame being
created. Once the function is done being executing we restore the stack
pointer to the value of the base pointer and resume executing from where
we left.