Fix evaluator tests.
The right way to create constants in C is to declare them as extern in
the header and initialize them in the C file. Fix this for the monkey
constant objects (true, false and null). This fixes the tests for
evaluator.

Also simplify the evaluator test for hash literals.
Simplify the code for evaluating hash literals.