Overhaul the object framework to avoid execessive memory allocations
Instead of copying and freeing monkey objects, we can maintain some sort
of a reference count to each object and decrement each time a reference
is freed. This should reduce the need to do memory allocations and
freeing up of memory. This improves the performance on the benchemark
significantly.

The performance has come down to about 6.3 seconds from 40 seconds with
the VM (with compiler optimizations enabled).